+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## 2.0.0 (TBD)
 
 - Minimum supported Symfony version is 4.4
+- Removed deprecated service aliases
+- Removed reserved router name list